+++ src/LoginFormContainer.jsx
@@ -32,7 +32,7 @@ export default function LoginFormContainer() {
   };
 
   return (
-    <>
+    <div>
       {accessToken ? (
         <LogoutForm onClick={handleClickLogout} />
       ) : (
@@ -42,6 +42,6 @@ export default function LoginFormContainer() {
           onSubmit={handleSubmit}
         />
       )}
-    </>
+    </div>
   );
 }
